Code refactor: Change package of video convertion classes
