+++ bbb-video/src/main/java/org/bigbluebutton/app/video/VideoApplication.java
@@ -22,7 +22,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
-import org.bigbluebutton.app.video.h263.H263Converter;
+import org.bigbluebutton.app.video.converter.H263Converter;
 import org.red5.logging.Red5LoggerFactory;
 import org.red5.server.adapter.MultiThreadedApplicationAdapter;
 import org.red5.server.api.IConnection;
