+++ bbb-video/src/main/java/org/bigbluebutton/app/video/ffmpeg/ProcessMonitor.java
@@ -1,4 +1,4 @@
-package org.bigbluebutton.app.video.h263;
+package org.bigbluebutton.app.video.ffmpeg;
 
 import java.io.InputStream;
